requirements.txt
python-gnupg>=0.4.7
requests>=2.31.0
pytest>=7.4.0
